ON-16121: Translate filter requests into mcdi commands
This commit changes the ef10ct filter_to_mcdi function into the more
generic ethtool_flow_to_mcdi_op which allows for the conversion between
an ethtool filter specification and an mcdi filter insert op.

This commit also updates some of the other handling for values return
from the filter insert op to make sure that they are properly stored in
onload. Additionally adds some basic filter checking in the test driver.

+++ src/include/ci/efhw/efct.h
@@ -192,7 +192,7 @@ static const size_t MAX_ALLOWED_mac_vlan = 128;
   action(mac_vlan)
 
 struct efct_hw_filter {
-  int drv_id;
+  uint64_t drv_id;
   unsigned refcount;
   uint32_t hw_id;
   uint8_t rxq;
